Speed up external dependencies lookup
Previously in order to find Analysis for given class name zinc fist have to enumerated whole classpath (each class in each jar or classes directory). In huge workspace it might take over 40 sec – and this was overhead for each compilation (even the smallest).

Now zinc searches for given analysis in all analysis ordered in the same way as classpath so whole classpath enumeration is skipped for searching for changed sources. Now overhead is barely  noticeable.

This might not detect correct changes if class x.X was previously present only in ouput of depended project but later class with same binary name was added to raw classes directory that is earlier on classpath. This is true if this situation happens when jar content is changed (so x.X is added there) however this should be capture as change in binary dependencies.

From reasons above I deiced to keep old behavior under flag.