Remove `compileScalaBridge` from util interface
The removed method in this commit should not be in the Zinc utils
provided by the public API because this is something that users of Zinc
should be handling themselves, choosing whichever strategy they want to
locate the JAR files containing the compiler interface and its
implementation.

For instance, Pants and SBT have different ways of doing this. We remove
this functionality from the public API to keep it simple and
understandable.

This commit also redefines the documentation of `compile` in the
`IncrementalCompiler` interface.