Create `ReadSourceInfos` & add to `CompileAnalysis`
`ReadSourceInfos` is a read-only interface for getting source info
(namely, compiler errors) associated with a given file.

This is very useful in lots of cases. Especially to process and improve
error messages in a lightweight way, though ScalaClippy already exists
for that purpose.

In any case, having access to the compiler errors outside of is of
ultimate importance for tools that want to preprocess compilation
results independently of sbt.