Add VirtualFile / VirtualFileRef / FileConverter
This implements Zinc support for passing source code as a VirtualFile datatype.
This allows build tools to pass source files as in-memory datatype.

Another motivation for doing so is to remove machine dependence from the internal state of incremental compilation (Analysis). By making Analysis free of absolute paths, we should be able to cache the file and resume compilation from another machine.